refactor: rearranged TS configs
