+++ .eslintrc.cjs
@@ -5,6 +5,7 @@ module.exports = {
   root: true,
   extends: [
     'plugin:vue/vue3-recommended',
+    'eslint:recommended',
     '@vue/eslint-config-typescript',
     '@vue/eslint-config-prettier/skip-formatting',
     'plugin:security/recommended',
