--- src/env.d.ts
@@ -1,34 +1,0 @@
-/// <reference types="vite/client" />
-import type { Router } from 'vue-router'
-import 'vue-router'
-import 'pinia'
-
-declare module '*.vue' {
-  import { DefineComponent } from 'vue'
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
-  const component: DefineComponent<any, {}, any>
-  export default component
-}
-
-interface ImportMetaEnv extends Readonly<Record<string, string>> {
-  // Only string type here to avoid hard to debug cast problems in your components!
-  readonly VITE_APP_VERSION: string
-  readonly VITE_APP_BUILD_EPOCH?: string
-}
-interface ImportMeta {
-  readonly env: ImportMetaEnv
-}
-
-declare module 'vue-router' {
-  interface RouteMeta {
-    title?: string
-  }
-}
-
-declare module 'pinia' {
-  export interface PiniaCustomProperties {
-    router: Router
-  }
-}
-
-export {}
