+++ tsconfig.node.json
@@ -1,12 +1,11 @@
 {
+  "extends": "@tsconfig/node18/tsconfig.json",
+  "include": ["vite.config.*", "vitest.config.*", "playwright.config.*"],
   "compilerOptions": {
     "composite": true,
-    "skipLibCheck": true,
+    "noEmit": true,
     "module": "ESNext",
-    "moduleResolution": "bundler",
-    "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true,
-    "types": ["node", "jsdom"]
-  },
-  "include": ["vite.config.ts", "vitest.config.ts", "playwright.config.ts", "package.json"]
+    "moduleResolution": "Bundler",
+    "types": ["node"]
+  }
 }
